% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmafster_calc.R
\name{RmafsterCalc}
\alias{RmafsterCalc}
\title{\lifecycle{experimental}}
\usage{
RmafsterCalc(mutations_tbl = NULL, samples_tbl = NULL)
}
\arguments{
\item{mutations_tbl}{a data frame of mutations.
Required columns:
chr: chromosome,
pos: genomic position,
ref: reference or wildtype allele,
alt: mutated allele.
Optional Columns:
symbol: gene symbol,
sample_id: the mutations are specific for samples, if \code{NULL} all mutations are searched in all samples,
var: mutations type, defaults to \code{'SNP'},
vaf: variant allele frequency for mutation (defaults to \code{0.5}),
dna_dp: depth of dna sequencing for mutation (defaults to \code{200})}

\item{samples_tbl}{a data frame of samples
Required columns:
sample_id: an id for each sample,
bam_path: full path for .bam file (.bai file must be present in the same directory).
Optional columns:
purity: tumour purity for sample (defaults to \code{1})}
}
\value{
a data frame copy of \code{mutations} with 6 additional columns: \code{"ref_alleles"}, \code{"alt_alleles"}, \code{"other_alleles"}, \code{"purity"}, \code{"rna_dp"} and \code{"rmaf"}.
}
\description{
calculates RMAFs for a set of mutations in specific BAM files.
}
\examples{
samples_tbl = data.frame(
               sample_id='CT26',
               bam_path=system.file("extdata","CT26_chr8_115305465.bam",
               package = 'RMAFster', mustWork=TRUE),
               purity=1,
               stringsAsFactors = FALSE)
mutations_tbl = data.frame(
                 chr='chr8',
                 pos=115305465,
                 ref='G',
                 alt='A',
                 sample_id='CT26',
                 symbol ='Cntnap4')
rmafs = RmafsterCalc(
     mutations_tbl,
     samples_tbl
)
}
