% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmafster_calc.R
\name{RmafsterCalc}
\alias{RmafsterCalc}
\title{\lifecycle{experimental}}
\usage{
RmafsterCalc(mutations = NULL, samples = NULL)
}
\arguments{
\item{mutations}{a data frame of mutations.
Required columns:
chr: chromosome,
pos: genomic position,
ref: reference or wildtype allele,
alt: mutated allele.
Optional Columns:
symbol: gene symbol,
sample_id: the mutations are specific for samples, otherwise all mutations will searched in all samples,
var: mutations type, defaults to 'SNP',
vaf: variant allele frequency for mutation (defaults to 0.5),
dna_dp: depth of dna sequencing for mutation (defaults to 200)}

\item{samples}{a data frame of samples
Required columns:
sample_id: an id for each sample,
bam_path: full path for bam file (bai must be present in directory).
Optional columns:
purity: tumour purity for sample (defaults to 1)}
}
\description{
calculates RMAFs for a set of mutations in specific BAM files.
}
\examples{
samples = data.frame(
               sample_id='CT26',
               bam_path=system.file("extdata","CT26_chr8_115305465.bam",
               package = 'RMAFster', mustWork=TRUE),
               purity=1,
               stringsAsFactors = FALSE)
mutations = data.frame(
                 chr='chr8',
                 pos=115305465,
                 ref='G',
                 alt='A',
                 sample_id='CT26',
                 symbol ='Cntnap4')
RmafsterCalc(
     mutations,
     samples
)
}
